# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Water app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of water app"
  },
  "servers": [
    {
      "url": "https://team-project-b-watter-app.onrender.com"
    },
    {
      "url": "http://localhost:14000"
    }
  ],
...
  "paths": {
...
    "/water": {
      "post": {
        "tags": [
          "water"
        ],
        "summary": "Add a new water consumption record",
        "operationId": "addWaterConsumption",
        "description": "Add a new record of water consumption for the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "string",
                    "example": "250"
                  },
                  "time": {
                    "type": "string",
                    "example": "00:10"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New water consumption record has been added successfully!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/waterPostAndPatchResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "get": {
        "tags": [
          "water"
        ],
        "summary": "Get water consumption in percent",
        "operationId": "getDailyWaterConsumption",
        "description": "Retrieve the amount of water consumed from the daily norm in percent and a list of all records of water consumption for the current day",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved daily water consumption data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dailyWaterConsumptionResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/water/{id}": {
      "patch": {
        "tags": [
          "water"
        ],
        "summary": "Update an existing water consumption record",
        "operationId": "updateWaterConsumption",
        "description": "Update the amount of water consumed in a specific record.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "66bb2c46d348911e38a4e9d5",
              "description": "The ID of the water consumption record to update."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "string",
                    "example": "250"
                  },
                  "time": {
                    "type": "string",
                    "example": "00:10"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Water consumption record has been updated successfully!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/waterPostAndPatchResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "tags": [
          "water"
        ],
        "summary": "Delete a water consumption record",
        "operationId": "deleteWaterConsumption",
        "description": "Remove a water consumption record by ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "66bb2c46d348911e38a4e9d5",
              "description": "The ID of the water consumption record to delete."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Water consumption record has been deleted successfully!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/waterDeleteResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/water/{month}/{year}": {
      "get": {
        "tags": [
          "water"
        ],
        "summary": "Get water consumption by month",
        "operationId": "getWaterConsumptionByMonth",
        "description": "Retrieve the water consumption records for each day of a selected month for the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number",
              "example": 8
            },
            "description": "The month to retrieve records for."
          },
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number",
              "example": 2024
            },
            "description": "The year to retrieve records for."
          }
        ],
        "responses": {
          "200": {
            "description": "Month period records of water consumption per day for userId: 66d061dc16f361d7057ce7ca",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/waterGetByMonthResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "dailyWaterConsumptionResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Successfully retrieved daily water consumption data"
          },
          "data": {
            "type": "object",
            "required": [
              "persantageOfNorm",
              "dailyRecords",
              "createdAt",
              "updatedAt"
            ]
          }
        }
      },
      "waterPostAndPatchResponse": {
        "type": "object",
        "required": [
          "_id",
          "userId",
          "date",
          "amount",
          "dailyNorm",
          "percentageConsumed",
          "consumedWaterByDay",
          "entries",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "description": "Record's id\"",
            "type": "string",
            "example": "66d4347c5e230c3b03bff562"
          },
          "userId": {
            "description": "Record's user id",
            "type": "string",
            "example": "66d061dc16f361d7057ce7ca"
          },
          "date": {
            "description": "Date of adding the record `dd, MonthName` format",
            "type": "string",
            "example": "01, September"
          },
          "time": {
            "description": "Time of the record creation or updating in `hh:mm` format",
            "type": "string",
            "example": "00:10"
          },
          "amount": {
            "description": "Amount of water consumed by user added to the record, in milliliters",
            "type": "number",
            "example": 250
          },
          "dailyNorm": {
            "description": "User's daily water intake norm, in litres",
            "type": "number",
            "example": "2.0 L"
          },
          "percentageConsumed": {
            "description": "Percentage of the daily water intake norm consumed",
            "type": "string",
            "example": "50%"
          },
          "consumedWaterByDay": {
            "description": "Total amount of water consumed by the user for the day, in litres",
            "type": "string",
            "example": "1.0 L"
          },
          "entries": {
            "type": "number",
            "example": 6,
            "description": "The number of water consumption entries by day."
          },
          "createdAt": {
            "description": "Timestamp when the record was created",
            "type": "string",
            "format": "date-time",
            "example": "2024-08-28T14:48:00.000Z"
          },
          "updatedAt": {
            "description": "Timestamp when the record was last updated",
            "type": "string",
            "format": "date-time",
            "example": "2024-08-28T14:55:00.000Z"
          }
        }
      },
      "waterDeleteResponse": {
        "type": "object",
        "required": [
          "_id",
          "userId",
          "date",
          "amount",
          "dailyNorm",
          "percentageConsumed",
          "consumedWaterByDay",
          "entries"
        ],
        "properties": {
          "_id": {
            "description": "Deleted record's id\"",
            "type": "string",
            "example": "66d4347c5e230c3b03bff562"
          },
          "userId": {
            "description": "Record's user id",
            "type": "string",
            "example": "66d061dc16f361d7057ce7ca"
          },
          "date": {
            "description": "Date of deleting the record `dd, MonthName` format",
            "type": "string",
            "example": "01, September"
          },
          "time": {
            "description": "Time of the record deleting in `hh:mm` format",
            "type": "string",
            "example": "00:10"
          },
          "amount": {
            "description": "Amount of water consumed by user of deleted record, in milliliters",
            "type": "number",
            "example": 250
          },
          "dailyNorm": {
            "description": "User's daily water intake norm, in litres",
            "type": "number",
            "example": "2.0 L"
          },
          "percentageConsumed": {
            "description": "Updated Percentage of the daily water intake norm consumed after deleting the record",
            "type": "string",
            "example": "50%"
          },
          "consumedWaterByDay": {
            "description": "Updated total amount of water consumed by the user for the day after deleting the record, in litres",
            "type": "string",
            "example": "1.0 L"
          },
          "entries": {
            "type": "number",
            "example": 6,
            "description": "The number of water consumption entries by day."
          },
          "createdAt": {
            "description": "Timestamp when the record was created",
            "type": "string",
            "format": "date-time",
            "example": "2024-08-28T14:48:00.000Z"
          },
          "updatedAt": {
            "description": "Timestamp when the record was last updated",
            "type": "string",
            "format": "date-time",
            "example": "2024-08-28T14:55:00.000Z"
          }
        }
      },
      "waterGetByMonthResponse": {
        "type": "array",
        "items": {
          "type": "object",
          "required": [
            "date",
            "dailyNorm",
            "percentageConsumed",
            "entries"
          ],
          "properties": {
            "date": {
              "type": "string",
              "example": "01, September",
              "description": "The date in `dd, MonthName` format."
            },
            "dailyNorm": {
              "type": "string",
              "example": "2,0 L",
              "description": "The daily water intake norm in liters."
            },
            "percentageConsumed": {
              "type": "string",
              "example": "64%",
              "description": "The percentage of the daily norm consumed."
            },
            "entries": {
              "type": "number",
              "example": 6,
              "description": "The number of water consumption entries."
            }
          }
        }
      }
    },
    "responses": {
...
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Not found"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
    }
  }
}